safe robot moves | chandan | 10th April
An island is represented in an 2d array of size mxn. If robot moves out
of island it gets
destroyed and robot is allowed to move top, down, left, right. Given a
robot position (x, y)
and k steps, what is the probability of robot getting survived in the
effort of making
k steps.